server/requirements.txt
alabaster==0.7.12
alembic==1.2.1
altgraph==0.16.1
anaconda==0.0.1.1
asn1crypto==0.24.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
blinker==1.4
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.4.1
cryptography==2.3.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
docutils==0.15.2
editdistance==0.5.3
entrypoints==0.3
et-xmlfile==1.0.1
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
future==0.17.1
futures==3.1.1
get==2019.4.13
gitdb2==2.0.4
GitPython==2.1.11
goslate==1.5.1
gunicorn==19.8.1
idna==2.6
imagesize==1.1.0
ipykernel==5.1.2
ipyparallel==6.2.4
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==0.24
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10
joblib==0.13.2
jsonschema==3.0.2
jupyter-client==5.3.2
jupyter-core==4.5.0
kiwisolver==1.1.0
lxml==4.4.1
Mako==1.1.0
MarkupSafe==1.0
matplotlib==3.1.1
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
ndg-httpsclient==0.5.1
nose==1.3.7
notebook==6.0.1
numpy==1.16.1
openpyxl==2.6.2
packaging==19.1
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==5.1.0
pipenv==2018.11.26
post==2019.4.13
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psycopg2-binary==2.7.5
public==2019.4.13
pusher==2.0.1
pyasn1==0.4.4
pycparser==2.19
PyDictionary==1.5.2
pygame==1.9.4
Pygments==2.4.2
PyInstaller==3.5
PyJWT==1.7.1
pyOpenSSL==18.0.0
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
pywin32==225
pywin32-ctypes==0.2.0
pywinpty==0.5.5
pyzmq==18.1.0
qtconsole==4.5.5
query-string==2019.4.13
request==2019.4.13
requests==2.18.4
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.11.0
sklearn==0.0
smmap2==2.0.4
snowballstemmer==1.9.1
soupsieve==1.9.4
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.10
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.22
virtualenv==16.0.0
virtualenv-clone==0.4.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.5.1
WTForms==2.2.1
